t-index/requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-cors==0.8.1
aiosignal==1.4.0
airportsdata==20260315
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astor==0.8.1
attrs==26.1.0
bitsandbytes==0.49.2
blake3==1.0.8
build==1.4.1
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
cmake==4.3.0
colorful==0.5.8
compressed-tensors==0.9.3
cryptography==46.0.5
cuda-pathfinder==1.5.0
cupy-cuda12x==14.0.1
datasets==4.8.4
deepspeed==0.17.1
deprecated==1.3.1
depyf==0.18.0
dill==0.4.1
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
einops==0.8.2
email-validator==2.3.0
evaluate==0.4.6
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.9.0
filelock==3.25.2
flash-attn==2.7.4.post1
frozenlist==1.8.0
fsspec==2026.2.0
gguf==0.18.0
gitdb==4.0.12
gitpython==3.1.46
google-api-core==2.30.0
google-auth==2.49.1
googleapis-common-protos==1.73.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.2
hjson==3.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==1.7.2
idna==3.11
importlib-metadata==8.0.0
interegular==0.3.3
isort==8.0.1
jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonlines==4.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lark==1.2.2
liger-kernel==0.7.0
lightning-utilities==0.15.3
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
loralib==0.1.2
markdown==3.10.2
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
mistral-common==1.10.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.1
multiprocess==0.70.19
nest-asyncio==1.6.0
networkx==3.6.1
ninja==1.13.0
numba==0.61.2
numpy==2.4.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==13.595.45
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvtx==0.2.15
openai==2.29.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.13.0.92
openrlhf==0.8.4
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.13
optimum==2.1.0
optree==0.19.0
outlines==0.1.11
outlines-core==0.1.26
packaging==26.0
pandas==3.0.1
partial-json-parser==0.2.1.1.post7
peft==0.18.1
pillow==12.1.1
pip @ file:///home/task_177071101768311/croot/pip_1770711076895/work
platformdirs==4.9.4
prometheus-client==0.24.1
prometheus-fastapi-instrumentator==7.1.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==4.25.8
psutil==7.2.2
py-cpuinfo==9.0.0
py-spy==0.4.1
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycountry==26.2.16
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pygments==2.19.2
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-discovery==1.2.0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-multipart==0.0.22
pyyaml==6.0.3
pyzmq==27.1.0
ray==2.43.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
ring-flash-attn==0.1.8
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.1
sentry-sdk==2.55.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart-open==7.5.1
smmap==5.0.3
sniffio==1.3.1
starlette==0.52.1
sympy==1.13.1
tabulate==0.10.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchdata==0.11.0
torchmetrics==1.9.0
torchvision==0.21.0+cu124
tqdm==4.67.3
transformers==5.3.0
transformers-stream-generator==0.0.5
triton==3.2.0
trl==0.19.1
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uv==0.10.12
uvicorn==0.42.0
uvloop==0.22.1
virtualenv==21.2.0
vllm==0.8.5.post1
wandb==0.25.1
watchfiles==1.1.1
websockets==16.0
werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.2
xformers==0.0.29.post2
xgrammar==0.1.18
xielu @ git+https://github.com/nickjbrowning/XIELU@2a55f6b9efa64954bf173e63297a2b2f99741b69
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
